// packages/sdks/gatekeeper/src/lib/types.ts
export interface GrantFilter {
  readonly principalType?: PrincipalType | undefined;
  readonly principalId?: string | undefined;
  readonly resourceType?: string | undefined;
  readonly resourceId?: string | undefined;
  readonly used?: boolean | undefined;
}